refactor: extract theme toggle script
